fix: Update setup
