+++ modules/docs/mdx/16.0-UPGRADE-GUIDE.mdx
@@ -75,8 +75,9 @@ against `@workday/canvas-tokens-web/css/sana/_variables.css`:
 
 The `CanvasProvider` theming updates in
 [#4060](https://github.com/Workday/canvas-kit/pull/4060) are for **scoped** use cases — embedding
-Canvas in another brand, multi-tenant sections, and popup parity. Application teams should not need
-them if they import the Sana variables and set `data-theme="sana-canvas"` globally.
+Canvas in another brand, multi-tenant sections, and popup parity when you cannot set
+`data-theme="sana-canvas"` on `<html>`. Application teams that control `<html>` and import the
+Sana variables globally do not need a `theme` prop for app-wide Sana.
 
 For the scoped theming API, see our
 [Theming documentation](https://workday.github.io/canvas-kit/?path=/docs/features-theming-overview--docs).
@@ -109,9 +110,8 @@ relying on the auto-generated color ramp need to explicitly opt into `'full'` sc
 
 ### Simplified Sana Canvas Setup
 
-If you're using Sana Canvas globally with `data-theme="sana-canvas"`, you no longer need to pass
-`sanaCanvasProviderTheme` to CanvasProvider. Popups naturally inherit CSS variables from the
-global theme.
+If you can set `data-theme="sana-canvas"` on `<html>`, you no longer need to pass
+`sanaCanvasProviderTheme` to CanvasProvider. Popups inherit CSS variables from the document.
 
 **Before:**
 ```tsx
@@ -122,20 +122,30 @@ import {CanvasProvider, sanaCanvasProviderTheme} from '@workday/canvas-kit-react
 </CanvasProvider>
 ```
 
-**After:**
+**After (when you control `<html>`):**
 ```tsx
 import {CanvasProvider} from '@workday/canvas-kit-react/common';
 
-// Simpler setup - no theme prop needed
+// <html data-theme="sana-canvas">
 <CanvasProvider>
   <App />
 </CanvasProvider>
 ```
 
-The `sanaCanvasProviderTheme` export remains available for edge cases like testing environments
-without global CSS or custom popup containers, but is no longer required for typical usage.
+**Still required — no access to `<html>`:** embedded apps, microfrontends, and third-party shells
+often cannot set attributes on `<html>`. Nested `data-theme` does not reach portaled popups. Keep
+passing `sanaCanvasProviderTheme` so menus, selects, and modals get Sana brand variables:
 
-If you see a console warning about unnecessary theme usage, you can safely remove the theme prop.
+```tsx
+import {CanvasProvider, sanaCanvasProviderTheme} from '@workday/canvas-kit-react/common';
+
+<CanvasProvider theme={sanaCanvasProviderTheme}>
+  <App />
+</CanvasProvider>
+```
+
+If you control `<html>` and see a console warning about unnecessary theme usage, you can safely
+remove the theme prop.
 
 ## Table of Contents
