+++ modules/react/common/stories/mdx/examples/SimplifiedSanaSetup.tsx
@@ -1,0 +1,30 @@
+import React from 'react';
+
+import {PrimaryButton, SecondaryButton} from '@workday/canvas-kit-react/button';
+import {CanvasProvider, sanaCanvasProviderTheme} from '@workday/canvas-kit-react/common';
+import {Menu} from '@workday/canvas-kit-react/menu';
+import {Popup, useCloseOnOutsideClick, usePopupModel} from '@workday/canvas-kit-react/popup';
+
+export const SimplifiedSetup = () => {
+  const myModel = usePopupModel();
+  useCloseOnOutsideClick(myModel);
+  return (
+    <CanvasProvider theme={sanaCanvasProviderTheme} data-theme="sana-canvas">
+      <Popup model={myModel}>
+        <Popup.Target as={SecondaryButton}>Open Menu</Popup.Target>
+        <Popup.Popper>
+          <Popup.Card>
+            <Popup.Body>
+              <Menu>
+                <Menu.Item>Option 1</Menu.Item>
+                <Menu.Item>Option 2</Menu.Item>
+                <Menu.Item>Option 3</Menu.Item>
+              </Menu>
+              <PrimaryButton>Hello World</PrimaryButton>
+            </Popup.Body>
+          </Popup.Card>
+        </Popup.Popper>
+      </Popup>
+    </CanvasProvider>
+  );
+};
